Cleaning up code ordering in some files, adding more docs, and ensuring that CameraLabelerDrawer can be derived properly in user code.
