+++ com.unity.perception/Editor/GroundTruth/CameraLabelerDrawer.cs.meta
@@ -1,0 +1,3 @@
+﻿fileFormatVersion: 2
+guid: a9aa79de1f2f402aa15ef42f82c3d334
+timeCreated: 1593613195
