+++ com.unity.perception/Editor/GroundTruth/CameraLabelerDrawerAttribute.cs.meta
@@ -1,0 +1,3 @@
+﻿fileFormatVersion: 2
+guid: b7dfd54d7de84ca6a06e36f0e09de906
+timeCreated: 1593613460
