+++ com.unity.perception/Editor/GroundTruth/LabelingEditor.cs
@@ -12,7 +12,7 @@ namespace UnityEditor.Perception.GroundTruth
 
         public void OnEnable()
         {
-            m_LabelsList = new ReorderableList(serializedObject, serializedObject.FindProperty(nameof(global::UnityEngine.Perception.GroundTruth.Labeling.labels)), true, false, true, true);
+            m_LabelsList = new ReorderableList(serializedObject, serializedObject.FindProperty(nameof(Labeling.labels)), true, false, true, true);
             m_LabelsList.drawElementCallback = DrawElement;
             m_LabelsList.onAddCallback += OnAdd;
             m_LabelsList.onRemoveCallback += OnRemove;
@@ -21,14 +21,14 @@ namespace UnityEditor.Perception.GroundTruth
         void OnRemove(ReorderableList list)
         {
             if (list.index != -1)
-                Labeling.labels.RemoveAt(list.index);
+                labeling.labels.RemoveAt(list.index);
         }
 
-        Labeling Labeling => (Labeling)target;
+        Labeling labeling => (Labeling)target;
 
         void OnAdd(ReorderableList list)
         {
-            Labeling.labels.Add("");
+            labeling.labels.Add("");
         }
 
         void DrawElement(Rect rect, int index, bool isactive, bool isfocused)
@@ -40,10 +40,10 @@ namespace UnityEditor.Perception.GroundTruth
                     return;
 
                 var contentRect = new Rect(rect.x + indent, rect.y, rect.width - indent, rect.height);
-                var value = EditorGUI.TextField(contentRect, Labeling.labels[index]);
+                var value = EditorGUI.TextField(contentRect, labeling.labels[index]);
                 if (change.changed)
                 {
-                    Labeling.labels[index] = value;
+                    labeling.labels[index] = value;
                 }
             }
         }
