+++ com.unity.perception/Runtime/GroundTruth/Labeling/SemanticSegmentationLabelConfig.cs
@@ -9,11 +9,20 @@ namespace UnityEngine.Perception.GroundTruth {
     {
     }
 
+    /// <summary>
+    /// LabelEntry for <see cref="SemanticSegmentationLabelConfig"/>. Maps a label to a color.
+    /// </summary>
     [Serializable]
     public struct SemanticSegmentationLabelEntry : ILabelEntry
     {
         string ILabelEntry.label => this.label;
+        /// <summary>
+        /// The label this entry should match.
+        /// </summary>
         public string label;
+        /// <summary>
+        /// The color to be drawn in the semantic segmentation image
+        /// </summary>
         public Color color;
     }
 }
